mhc: consolidate device modules and tests, drop multichip cases
Folds mhc_ttnn/mhc_block/mhc_kernel into one tt_mhc.py that mirrors the torch
reference class for class, dropping the composite parametrization now that the
fused kernel is the only path. The kernel op is per-token with no cross-device
or hidden-dim dependency, so the multichip cases only exercised the mesh
harness; op coverage moves to the single-chip l2 nightly and the composite to
pcc/test_mhc.py. Worst PCC vs fp32 torch is 0.99884 on the C=7168 projection
matmul (its fp32 ceiling at reduction depth 28672); all mHC math is >= 0.99998.
