--- models/demos/deepseek_v3_d_p/tests/pcc/test_mhc_ttnn_pcc.py
@@ -1,194 +1,0 @@
-# SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
-# SPDX-License-Identifier: Apache-2.0
-
-"""PCC: TTNN composite mHC (models/demos/deepseek_v3_d_p/tt/mhc/mhc_ttnn.py) vs the pure-torch
-ground truth (models/demos/deepseek_v3_d_p/reference/mhc/mhc_reference.py).
-
-Each mHC piece is checked independently so a failure localises immediately:
-    project | parametrize (Sinkhorn) | hc_pre | hc_post | hc_head | full block.
-
-Precision note. Everything mHC-specific -- the Sinkhorn parametrization and the stream
-mixing -- matches fp32 torch to ~1e-7 PCC. The one lossy step is the projection matmul
-mixes = RMSNorm(X) @ P: at the model width n*C = 28672 the TT fp32 matmul tops out at
-~0.9989 PCC vs fp32 torch (HiFi4 + fp32 accumulation is already max fidelity -- this is
-the hardware, not this code). In the real model that error is squashed by the sigmoid and
-the small init a_res~0.01, so end-to-end pieces are checked at that scale; the Sinkhorn's
-own robustness to large logits is proven separately in test_parametrize at a_res up to 1.0.
-"""
-
-import os
-import sys
-
-import pytest
-import torch
-import torch.nn.functional as F
-from loguru import logger
-
-import ttnn
-
-sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "..", "reference", "mhc"))
-from mhc_reference import MHCConfig, MHCHead, MHCWrap, parametrize  # noqa: E402
-
-from models.demos.deepseek_v3_d_p.tt.mhc.mhc_ttnn import TtMHC, TtMHCHead  # noqa: E402
-
-PCC = 0.999
-PCC_PROJ = 0.998  # TT fp32-matmul ceiling at reduction depth n*C=28672 (see module docstring)
-
-
-def _cfg(C):
-    return MHCConfig(dim=C, n=4, sinkhorn_iters=20)
-
-
-def _params(cfg, scale_val, seed):
-    """Shared trainable params for reference and device."""
-    g = torch.Generator().manual_seed(seed)
-    fn = torch.randn(cfg.mix_hc, cfg.n * cfg.dim, generator=g) * 0.02
-    base = torch.randn(cfg.mix_hc, generator=g)  # non-zero to exercise the biases
-    scale = torch.full((3,), float(scale_val))
-    return fn, base, scale
-
-
-def _up(device, t):
-    return ttnn.from_torch(t.float(), layout=ttnn.TILE_LAYOUT, device=device, dtype=ttnn.float32)
-
-
-@pytest.mark.parametrize("T", [1, 32], ids=["T1", "T32"])
-@pytest.mark.parametrize("C", [256, 7168], ids=["C256", "C7168"])
-@pytest.mark.parametrize("scale_val", [0.01, 1.0], ids=["s0.01", "s1.0"])
-def test_parametrize(device, T, C, scale_val):
-    torch.manual_seed(0)
-    cfg = _cfg(C)
-    fn, base, scale = _params(cfg, scale_val, seed=1)
-    mixes = torch.randn(1, 1, T, cfg.mix_hc)
-
-    r_pre, r_post, r_comb = parametrize(mixes, scale, base, cfg, constraint="sinkhorn")
-
-    mhc = TtMHC(device, cfg, fn, base, scale)
-    d_pre, d_post, d_comb = mhc.parametrize(_up(device, mixes))
-
-    _check("pre", r_pre, ttnn.to_torch(d_pre))
-    _check("post", r_post, ttnn.to_torch(d_post))
-    _check("comb", r_comb, ttnn.to_torch(d_comb))
-
-
-@pytest.mark.parametrize("T", [1, 32], ids=["T1", "T32"])
-@pytest.mark.parametrize("C", [256, 7168], ids=["C256", "C7168"])
-def test_project(device, T, C):
-    torch.manual_seed(0)
-    cfg = _cfg(C)
-    fn, base, scale = _params(cfg, 1.0, seed=2)
-    x = torch.randn(1, T, cfg.n, C)
-
-    xf = x.reshape(T, cfg.n * C).float()
-    rsqrt = torch.rsqrt(xf.square().mean(-1, keepdim=True) + cfg.norm_eps)
-    r_mixes = (F.linear(xf, fn) * rsqrt).reshape(1, 1, T, cfg.mix_hc)
-
-    mhc = TtMHC(device, cfg, fn, base, scale)
-    d_mixes = mhc.project(_up(device, x))
-    _check("mixes", r_mixes, ttnn.to_torch(d_mixes), PCC_PROJ)
-
-
-# C256 uses aggressive a_res=1.0 (projection is precise at this width); C7168 uses the
-# model's init a_res~0.01 -- isolates the projection-matmul precision variable (docstring).
-_E2E = [(256, 1.0), (7168, 0.01)]
-_E2E_IDS = ["C256-s1.0", "C7168-s0.01"]
-
-
-@pytest.mark.parametrize("T", [1, 32], ids=["T1", "T32"])
-@pytest.mark.parametrize("C, scale_val", _E2E, ids=_E2E_IDS)
-def test_hc_pre(device, T, C, scale_val):
-    torch.manual_seed(0)
-    cfg = _cfg(C)
-    fn, base, scale = _params(cfg, scale_val, seed=3)
-    x = torch.randn(1, T, cfg.n, C)
-
-    wrap = _ref_wrap(cfg, fn, base, scale)
-    r_y, r_post, r_comb = wrap.hc_pre(x)  # [1,T,C], [1,T,n], [1,T,n,n]
-
-    mhc = TtMHC(device, cfg, fn, base, scale)
-    d_y, d_post, d_comb = mhc.hc_pre(_up(device, x))
-    _check("y", r_y, ttnn.to_torch(d_y))
-    _check("post", r_post, ttnn.to_torch(d_post))
-    _check("comb", r_comb, ttnn.to_torch(d_comb))
-
-
-@pytest.mark.parametrize("T", [1, 32], ids=["T1", "T32"])
-@pytest.mark.parametrize("C", [256, 7168], ids=["C256", "C7168"])
-def test_hc_post(device, T, C):
-    torch.manual_seed(0)
-    cfg = _cfg(C)
-    fn, base, scale = _params(cfg, 1.0, seed=4)
-    x = torch.randn(1, T, cfg.n, C)
-    f_out = torch.randn(1, T, C)
-    post = 2 * torch.sigmoid(torch.randn(1, T, cfg.n))
-    from mhc_reference import sinkhorn_knopp
-
-    comb = sinkhorn_knopp(torch.randn(1, T, cfg.n, cfg.n), cfg.sinkhorn_iters, cfg.eps)
-
-    wrap = _ref_wrap(cfg, fn, base, scale)
-    r_out = wrap.hc_post(f_out, x, post, comb)  # [1,T,n,C]
-
-    mhc = TtMHC(device, cfg, fn, base, scale)
-    d_out = mhc.hc_post(
-        _up(device, f_out.reshape(1, T, 1, C)),
-        _up(device, x),
-        _up(device, post.reshape(1, 1, T, cfg.n)),
-        _up(device, comb.reshape(1, T, cfg.n, cfg.n)),
-    )
-    _check("out", r_out, ttnn.to_torch(d_out))
-
-
-@pytest.mark.parametrize("T", [1, 32], ids=["T1", "T32"])
-@pytest.mark.parametrize("C", [256, 7168], ids=["C256", "C7168"])
-def test_hc_head(device, T, C):
-    torch.manual_seed(0)
-    cfg = _cfg(C)
-    g = torch.Generator().manual_seed(5)
-    fn = torch.randn(cfg.n, cfg.n * C, generator=g) * 0.02
-    base = torch.randn(cfg.n, generator=g)
-    scale = torch.full((1,), 0.01)
-    x = torch.randn(1, T, cfg.n, C)
-
-    head = MHCHead(cfg)
-    head.fn.data, head.base.data, head.scale.data = fn, base, scale
-    r_y = head(x)  # [1,T,C]
-
-    d_head = TtMHCHead(device, cfg, fn, base, scale)
-    d_y = d_head(_up(device, x))
-    _check("head_y", r_y, ttnn.to_torch(d_y))
-
-
-@pytest.mark.parametrize("T", [1, 32], ids=["T1", "T32"])
-@pytest.mark.parametrize("C, scale_val", _E2E, ids=_E2E_IDS)
-def test_full_block(device, T, C, scale_val):
-    """expand -> hc_pre -> F(identity) -> hc_post, device vs reference."""
-    torch.manual_seed(0)
-    cfg = _cfg(C)
-    fn, base, scale = _params(cfg, scale_val, seed=6)
-    x = torch.randn(1, T, cfg.n, C)
-
-    wrap = _ref_wrap(cfg, fn, base, scale)
-    r_out = wrap(x, lambda z: z)  # identity sublayer -> isolates mHC
-
-    mhc = TtMHC(device, cfg, fn, base, scale)
-    x_tt = _up(device, x)
-    y, post, comb = mhc.hc_pre(x_tt)
-    d_out = mhc.hc_post(y, x_tt, post, comb)  # F = identity
-    _check("block_out", r_out, ttnn.to_torch(d_out))
-
-
-def _ref_wrap(cfg, fn, base, scale):
-    wrap = MHCWrap(cfg, constraint="sinkhorn")
-    wrap.fn.data, wrap.base.data, wrap.scale.data = fn, base, scale
-    return wrap
-
-
-def _check(name, ref, dev, pcc=PCC):
-    ref = ref.float().flatten()
-    dev = dev.float().flatten()
-    md = (ref - dev).abs().max().item()
-    from models.common.utility_functions import comp_pcc
-
-    passed, msg = comp_pcc(ref, dev, pcc)
-    logger.info(f"{name}: pcc={msg} | max|Δ|={md:.2e}")
-    assert passed, f"{name}: pcc={msg} | max|Δ|={md:.2e} (threshold {pcc})"
